refactor: rename tenor_utils.py to tenor.py
